requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
reportlab>=4.0
python-dateutil>=2.8
flask-bcrypt
waitress
